requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
psycopg2-binary
bcrypt
numpy
lightfm
scipy
pandas